backend/api/filters.py
import django_filters
from .models import Deal

class DealFilter(django_filters.FilterSet):
    price_min = django_filters.NumberFilter(field_name='price', lookup_expr='gte')
    price_max = django_filters.NumberFilter(field_name='price', lookup_expr='lte')
    score_min = django_filters.NumberFilter(field_name='score', lookup_expr='gte')
    score_max = django_filters.NumberFilter(field_name='score', lookup_expr='lte')

    class Meta:
        model = Deal
        fields = ['price_min', 'price_max', 'score_min', 'score_max']
