backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg[binary]==3.1.19
